agents/requirements.txt
google-adk
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
python-dotenv
httpx
